requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-decouple==2.1
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
Pyrogram==2.0.93
PySocks==1.7.1
requests==2.28.1
rsa==4.9
sqlparse==0.4.3
Telethon==1.26.1
TgCrypto==1.2.5
urllib3==1.26.13
